# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/user/editpwd": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "修改密码接口",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.editPasswordStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":20000,\"message\":\"修改成功\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取用户菜单列表以及昵称",
                "responses": {
                    "200": {
                        "description": "{\"code\":20000,\"message\":\"ok\",\"data\":\"\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户登录接口",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.registAndLoginStuct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":20000,\"message\":\"ok\",\"data\":{}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.editPasswordStruct": {
            "type": "object",
            "required": [
                "new_password",
                "new_password_again",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "description": "长度大于等于6,小于等于10,与旧密码不同",
                    "type": "string"
                },
                "new_password_again": {
                    "description": "长度大于等于6,小于等于10,与新密码相同",
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
        "api.registAndLoginStuct": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}